Dialogs: delay calling componentComplete until parent is set
The non-native dialogs are created by QQmlComponent::create(), and the
parent is set right after the dialogs are created. This would mean that
componentComplete would get called before the parent were set to
anything other than nullptr, and the ancestors could not be traversed in
order to find the nearest parent item, which is useful for finding the
window that the dialog belongs to.

Change-Id: I650a27a3d1fa85a204976e7a093b45e804324d45
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>